Initial launch bug
The first time you launch an app with this code you get a dialog box asking for permissions. If you don't click it immediately but wait just a second or so before clicking allow, you have to force quit and start the app to get position updates. I believe this fixes #6